+++ src/actions/index.js
@@ -1,0 +1,12 @@
+const CREATE_BOOK = 'CREATE_BOOK';
+const REMOVE_BOOK = 'REMOVE_BOOK';
+
+export const createBook = (book) => ({
+  type: CREATE_BOOK,
+  payload: book,
+});
+
+export const removeBook = (book) => ({
+  type: REMOVE_BOOK,
+  payload: book,
+});
